Enable Java 8 builds with Jabel
Enable Jabel for all Java builds, and switch the target release to 8. This allows Jabel to fix everything it can from Java 11+ syntax to Java 8 bytecode, but build breaks remain for APIs which are different.

Jabel docs: https://github.com/bsideup/jabel

Using Jabel v0.4.2, which was stable for a long time. There is a newer release I have not tried yet.