Fix Java 8 build errors in wpimath
* No Map.entry
* No isEmpty on Optional<>
* No List.of()
* @Deprecated has no args in Java 8.
* No canAccess in Reflection for method access checks in tests.
* Fall back to v0.41 of ejml, with a few API changes due to this.

The ejml version change is the big one here, and needs to be carefully evaluated. We were on 0.43.1, but they stopped providing Java 8 build artifacts a while ago. 0.41 is the last version with Java 8 build artifacts, so I used that for now.

We could see about building ejml from source for Java 8. Or maybe it doesn't matter for our use cases. I don't know at this time.